pyproject.toml
[project]
name = "weather-ml"
version = "0.1.0"
description = "Weather ML practice project — xarray, PyTorch Lightning, Hydra"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2",
    "lightning>=2.2",
    "xarray>=2024.1",
    "netcdf4>=1.6",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "matplotlib>=3.8",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "requests>=2.31",
    "earthkit-data>=0.10",
    "earthkit-data[cds]>=0.10",
    "zarr>=3.1.5",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "jupyter>=1.0",
    "ipykernel>=6.29",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/weather_ml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.uv]
environments = [
    "sys_platform == 'linux'",
    "sys_platform == 'darwin'",
]

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[dependency-groups]
dev = [
    "prek>=0.3.5",
]
